pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "macroa"
version = "0.1.0"
description = "Personal AI OS — kernel, drivers, skills"
requires-python = ">=3.11"

dependencies = [
    "openai>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "ruff", "mypy", "black"]

[project.scripts]
macroa = "macroa.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["macroa*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
